Initial parent and children refactoring
